Refactored DeconzConnection. Classes are now more object-oriented. Preparation for implementation of TestConnection in settings (fragment).
